Finished Skins for all the enemies
With these frames, we can rotate the enimies. These frames only cover rotation from facing up to facing left, but we can invert the plotting process to acheive other angles.

These frames also contain animation frames for the bugs that need it (which is all of them so far.)

Also, it includes the frames neccisary for animating bug explosions.

I'm not planning on making any more enemies at the moment, because this was a beating, but I'll probably still make some bosses.

My next project will be trying to find the original sounds and music from the arcade game.