Finished Power Up Emblems
Finished power up emblems for now.
Note that I did some fancy anti-aliasing stuff, so these files are a lot bigger than those for the bugs and ships. As a result, they will probably need to be left at a scale of 1. They also get grainy if you blow them up because of the anti-aliasing