Uploaded the function I used to test draw stuff with lots of documentation
This function also has the correct colors for stuff. PLEASE USE THESE COLORS :art:
I did lots of work to find the authentic rgb values.

Also note that I haven't actually tested this function header and the rest of the setup, although I know the function works inside my main for it. If you have any problems with it (if you actually use it) let me know and I'll fix it.